Refactored coref component
